requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-genai==1.1.0
httpx==0.28.1
pandas==2.2.3
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1
massive==2.4.0
